requirements.txt
# please sort by a-z
chromadb==0.4.15
falcon
gunicorn
numpy
openai>=1.3
pandas
pydantic==1.10.9
pysqlite3-binary
schema
scikit-learn==1.3.2
scipy==1.11.3
requests==2.31.0
beautifulsoup4==4.12.3
